refactor: implement review comments + move functions
