+++ lib/ts/recipe/emailverification/recipeImplementation.ts
@@ -3,7 +3,6 @@ import { Querier } from "../../querier";
 import NormalisedURLPath from "../../normalisedURLPath";
 import RecipeUserId from "../../recipeUserId";
 import { GetEmailForRecipeUserIdFunc, UserEmailInfo } from "./types";
-import type AccountLinkingRecipe from "../accountlinking/recipe";
 import { getUser } from "../..";
 import { UserContext } from "../../types";
 import { SessionContainerInterface } from "../session/types";
@@ -86,8 +85,8 @@ export default function getRecipeInterface(
                         if (emailInfo.status === "OK" && emailInfo.email === response.email) {
                             // we do this here to prevent cyclic dependencies.
                             // TODO: Fix this.
-                            let AccountLinking = require("../accountlinking/recipe").default.getInstance() as AccountLinkingRecipe;
-                            await AccountLinking.createPrimaryUserIdOrLinkByAccountInfoOrLinkToSessionIfProvided({
+                            let AuthUtils = require("../../authUtils").AuthUtils;
+                            await AuthUtils.linkToSessionIfProvidedElseCreatePrimaryUserIdOrLinkByAccountInfo({
                                 tenantId,
                                 inputUser: updatedUser,
                                 recipeUserId,
